+++ src/MusicStore.MVC/API/SongApiController.cs
@@ -1,4 +1,4 @@
-﻿using AutoMapper;
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -15,29 +15,30 @@ namespace MusicStore.MVC.API
 {
   [Route("api/Song")]
   [ApiController]
+  [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Roles = "Administrator")]
   public class SongApiController : ControllerBase
   {
     private readonly IUnitOfWork unitOfWork;
     private readonly ILogger logger;
     private readonly IAuthorizationService authorizationService;
     private readonly UserManager<User> userManager;
-    private readonly IMapper mapper;
 
     public SongApiController(IUnitOfWork unitOfWork,
       ILogger<SongApiController> logger,
       IAuthorizationService authorizationService,
-      UserManager<User> userManager,
-      IMapper mapper)
+      UserManager<User> userManager)
     {
       this.unitOfWork = unitOfWork;
       this.logger = logger;
       this.authorizationService = authorizationService;
       this.userManager = userManager;
-      this.mapper = mapper;
     }
 
     [HttpGet]
-    public async Task<ActionResult<PaggingResult<Song>>> GetSongsPage([FromQuery]PaggingQuery query)
+    [ProducesResponseType(200)]
+    [ProducesResponseType(500)]
+    [AllowAnonymous]
+    public async Task<ActionResult<IActionResult>> GetSongsPage([FromQuery]PaggingQuery query)
     {
       try
       {
@@ -53,7 +54,10 @@ namespace MusicStore.MVC.API
     }
 
     [HttpGet("{id}")]
-    public async Task<ActionResult<Song>> GetSong(int id)
+    [ProducesResponseType(200)]
+    [ProducesResponseType(404)]
+    [ProducesResponseType(500)]
+    public async Task<IActionResult> GetSong(int id)
     {
       try
       {
@@ -61,7 +65,7 @@ namespace MusicStore.MVC.API
         if (song == null)
           return NotFound();
 
-        return song;
+        return Ok(song);
       }
       catch (Exception ex)
       {
@@ -71,7 +75,10 @@ namespace MusicStore.MVC.API
     }
 
     [HttpPost]
-    public async Task<ActionResult<Song>> CreateSong([FromBody]SongForCreatingDto dto)
+    [ProducesResponseType(201)]
+    [ProducesResponseType(400)]
+    [ProducesResponseType(500)]
+    public async Task<IActionResult> CreateSong([FromBody]SongForCreatingDto dto)
     {
       if (!ModelState.IsValid)
         return BadRequest();
@@ -83,15 +90,15 @@ namespace MusicStore.MVC.API
           dto.AlbumId = null;
 
         // Set the owner of this song to the current signedIn user
-        //var currentUserId = userManager.GetUserId(User);
-        //dto.OwenerId = currentUserId;
+        var currentUserId = userManager.GetUserId(User);
+        dto.OwenerId = currentUserId;
 
         var songEntity = await unitOfWork.Songs.AddAsync(dto);
         if (!await unitOfWork.SaveAsync())
           throw new Exception("Creating song failed on save.");
 
         var song = await unitOfWork.Songs.GetAsync(songEntity.Id);
-        return song;
+        return StatusCode(201, song);
       }
       catch (Exception ex)
       {
@@ -101,7 +108,12 @@ namespace MusicStore.MVC.API
     }
 
     [HttpPut("{id}")]
-    public async Task<ActionResult<Song>> UpdateeSong(
+    [ProducesResponseType(200)]
+    [ProducesResponseType(400)]
+    [ProducesResponseType(401)]
+    [ProducesResponseType(404)]
+    [ProducesResponseType(500)]
+    public async Task<IActionResult> UpdateeSong(
       int id,
       [FromBody]SongForUpdatingDto dto)
     {
@@ -121,16 +133,16 @@ namespace MusicStore.MVC.API
         if (dto.AlbumId != null && !await unitOfWork.Albums.Exist(dto.AlbumId))
           dto.AlbumId = null;
 
-        //var isAuthorized = await authorizationService
-        //  .AuthorizeAsync(User, song.OwenerId, AutherazationOperations.OwenResourse);
-        //if (!isAuthorized.Succeeded)
-        //  return Unauthorized();
+        var isAuthorized = await authorizationService
+          .AuthorizeAsync(User, song.OwenerId, AutherazationOperations.OwenResourse);
+        if (!isAuthorized.Succeeded)
+          return Unauthorized();
 
         await unitOfWork.Songs.UpdateAsync(dto);
         await unitOfWork.SaveAsync();
 
         song = await unitOfWork.Songs.GetAsync(id);
-        return song;
+        return Ok(song);
       }
       catch (Exception ex)
       {
@@ -140,7 +152,11 @@ namespace MusicStore.MVC.API
     }
 
     [HttpDelete("{id}")]
-    public async Task<ActionResult> DeleteSong(int id)
+    [ProducesResponseType(204)]
+    [ProducesResponseType(401)]
+    [ProducesResponseType(404)]
+    [ProducesResponseType(500)]
+    public async Task<IActionResult> DeleteSong(int id)
     {
       try
       {
@@ -148,10 +164,10 @@ namespace MusicStore.MVC.API
         if (song == null)
           return NotFound();
 
-        //var isAuthorized = await authorizationService
-        //  .AuthorizeAsync(User, song.OwenerId, AutherazationOperations.OwenResourse);
-        //if (!isAuthorized.Succeeded)
-        //  return Unauthorized();
+        var isAuthorized = await authorizationService
+          .AuthorizeAsync(User, song.OwenerId, AutherazationOperations.OwenResourse);
+        if (!isAuthorized.Succeeded)
+          return Unauthorized();
 
         await unitOfWork.Songs.DeleteAsync(id);
         if (!await unitOfWork.SaveAsync())
